fix2

+++ IA/Rules/Grid.cs
@@ -55,7 +55,7 @@ namespace IA.Rules
         /// <param name="c"></param>
         /// <param name="quantity"></param>
         /// <returns>return true in a pawn has been updated</returns>
-        public bool SetQuantityInCoord(Coord c, float quantity)
+        public bool SetQuantityInCoord(Coord c, int quantity)
         {
             if (this._pawns.TryGetValue(c, out Pawn pawn))
             {
